mb/lenovo/t540p/devicetree: Implement proper backlight PWM config
Further backport the backlight-PWM handling from Skylake. Beside
configuring the PWM frequency in Hz, we also use the PCH's logic
for the brightness setting via BLM_PCH_OVERRIDE_ENABLE. Linux
would toggle it anyway and that might confuse our ASL code.

We assume that the 183Hz value that was set before for Slippy
variants was overridden by Linux with the 200Hz VBT value, like
it was for the Broadwell Chromebooks. So we set 200Hz for them
in the devicetrees. The calculated value for the T440p of 220Hz
seems sane and also matches the VBT.

Change-Id: I98f208005bb660ee7a98f6ff9335ca52e2a9c6be
Signed-off-by: Nico Huber <nico.h@gmx.de>
Signed-off-by: Peter Lemenkov <lemenkov@gmail.com>