mb/lenovo/t540p/dsdt: Move ACPI table support out of arch/x86 (3/5)
This change moves all ACPI table support in coreboot currently living
under arch/x86 into common code to make it architecture
independent. ACPI table generation is not really tied to any
architecture and hence it makes sense to move this to its own
directory.

In order to make it easier to review, this change is being split into
multiple CLs. This is change 3/5 which basically is generated by
running the following command:
$ git grep -iIl "arch/acpi" | xargs sed -i 's/arch\/acpi/acpi\/acpi/g'

BUG=b:155428745

Change-Id: I93b7d951537b62a134b0dd2f6785e5d2c96881bd
Signed-off-by: Furquan Shaikh <furquan@google.com>
Signed-off-by: Peter Lemenkov <lemenkov@gmail.com>