mb/lenovo/t540p/devicetree: convert panel delays to ms representation
For easier review of the switch to a new register struct in the
follow-up change, the panel delay times get converted from destination
register raw format to milliseconds representation in this change.

Formula for conversion of power cycle delay:

  gpu_panel_power_cycle_delay_ms =
    (gpu_panel_power_cycle_delay - 1) * 100

Formula for all others:

  gpu_panel_power_X_delay_ms = gpu_panel_power_X_delay / 10

The register names gain a suffix `_ms` and calculation of the
destination register raw values gets done in gma code now.

Change-Id: I5c4371aa93cc9da56727854f9bbd58c387d6294a
Signed-off-by: Michael Niewöhner <foss@mniewoehner.de>
Signed-off-by: Peter Lemenkov <lemenkov@gmail.com>